# cc: posix shell compiler wrappers -- take three