--- a/lib/spack/spack/test/cc.py
+++ b/lib/spack/spack/test/cc.py
@@ -13,13 +13,13 @@ import pytest
 
 from spack.paths import build_env_path
 from spack.util.environment import set_env, system_dirs
-from spack.util.executable import Executable
+from spack.util.executable import Executable, ProcessError
 
 #
 # Complicated compiler test command
 #
 test_args = [
-    '-I/test/include', '-L/test/lib', '-L/other/lib', '-I/other/include',
+    '-I/test/include', '-L/test/lib', '-L/with space/lib', '-I/other/include',
     'arg1',
     '-Wl,--start-group',
     'arg2',
@@ -31,7 +31,9 @@ test_args = [
     '-Xlinker', '-rpath', '-Xlinker', '/fourth/rpath',
     '-Wl,--rpath,/fifth/rpath', '-Wl,--rpath', '-Wl,/sixth/rpath',
     '-llib3', '-llib4',
-    'arg5', 'arg6']
+    'arg5', 'arg6',
+    '"-DDOUBLE_QUOTED_ARG"', "'-DSINGLE_QUOTED_ARG'",
+]
 
 #
 # Pieces of the test command above, as they should be parsed out.
@@ -43,7 +45,7 @@ test_include_paths = [
     '-I/test/include', '-I/other/include']
 
 test_library_paths = [
-    '-L/test/lib', '-L/other/lib']
+    '-L/test/lib', '-L/with space/lib']
 
 test_wl_rpaths = [
     '-Wl,-rpath,/first/rpath', '-Wl,-rpath,/second/rpath',
@@ -60,7 +62,9 @@ test_args_without_paths = [
     '-Wl,--start-group',
     'arg2', 'arg3', '-llib1', '-llib2', 'arg4',
     '-Wl,--end-group',
-    '-llib3', '-llib4', 'arg5', 'arg6']
+    '-llib3', '-llib4', 'arg5', 'arg6',
+    '"-DDOUBLE_QUOTED_ARG"', "'-DSINGLE_QUOTED_ARG'",
+]
 
 #: The prefix of the package being mock installed
 pkg_prefix = '/spack-test-prefix'
@@ -86,6 +90,17 @@ spack_ldlibs   = ['-lfoo']
 lheaderpad = ['-Wl,-headerpad_max_install_names']
 headerpad = ['-headerpad_max_install_names']
 
+target_args = ["-march=znver2", "-mtune=znver2"]
+
+# common compile arguments: includes, libs, -Wl linker args, other args
+common_compile_args = (
+    test_include_paths +
+    test_library_paths +
+    ['-Wl,--disable-new-dtags'] +
+    test_wl_rpaths +
+    test_args_without_paths
+)
+
 
 @pytest.fixture(scope='session')
 def wrapper_environment():
@@ -107,7 +122,7 @@ def wrapper_environment():
             SPACK_LINK_DIRS=None,
             SPACK_INCLUDE_DIRS=None,
             SPACK_RPATH_DIRS=None,
-            SPACK_TARGET_ARGS='',
+            SPACK_TARGET_ARGS="-march=znver2 -mtune=znver2",
             SPACK_LINKER_ARG='-Wl,',
             SPACK_DTAGS_TO_ADD='--disable-new-dtags',
             SPACK_DTAGS_TO_STRIP='--enable-new-dtags'):
@@ -124,9 +139,6 @@ def wrapper_flags():
             SPACK_LDFLAGS=' '.join(spack_ldflags),
             SPACK_LDLIBS=' '.join(spack_ldlibs)):
         yield
-
-
-pytestmark = pytest.mark.usefixtures('wrapper_environment')
 
 
 def check_args(cc, args, expected):
@@ -149,7 +161,7 @@ def check_env_var(executable, var, expected):
     """
     with set_env(SPACK_TEST_COMMAND='dump-env-' + var):
         output = executable(*test_args, output=str).strip()
-        assert output == executable.path + ': ' + var + ': ' + expected
+        assert executable.path + ': ' + var + ': ' + expected == output
 
 
 def dump_mode(cc, args):
@@ -158,7 +170,13 @@ def dump_mode(cc, args):
         return cc(*args, output=str).strip()
 
 
-def test_vcheck_mode():
+def test_no_wrapper_environment():
+    with pytest.raises(ProcessError):
+        output = cc(output=str)
+        assert "Spack compiler must be run from Spack" in output
+
+
+def test_vcheck_mode(wrapper_environment):
     assert dump_mode(cc, ['-I/include', '--version']) == 'vcheck'
     assert dump_mode(cc, ['-I/include', '-V']) == 'vcheck'
     assert dump_mode(cc, ['-I/include', '-v']) == 'vcheck'
@@ -167,17 +185,17 @@ def test_vcheck_mode():
     assert dump_mode(cc, ['-I/include', '-V', '-o', 'output']) == 'vcheck'
 
 
-def test_cpp_mode():
+def test_cpp_mode(wrapper_environment):
     assert dump_mode(cc, ['-E']) == 'cpp'
     assert dump_mode(cxx, ['-E']) == 'cpp'
     assert dump_mode(cpp, []) == 'cpp'
 
 
-def test_as_mode():
+def test_as_mode(wrapper_environment):
     assert dump_mode(cc, ['-S']) == 'as'
 
 
-def test_ccld_mode():
+def test_ccld_mode(wrapper_environment):
     assert dump_mode(cc, []) == 'ccld'
     assert dump_mode(cc, ['foo.c', '-o', 'foo']) == 'ccld'
     assert dump_mode(cc, ['foo.c', '-o', 'foo', '-Wl,-rpath,foo']) == 'ccld'
@@ -185,13 +203,13 @@ def test_ccld_mode():
         'foo.o', 'bar.o', 'baz.o', '-o', 'foo', '-Wl,-rpath,foo']) == 'ccld'
 
 
-def test_ld_mode():
+def test_ld_mode(wrapper_environment):
     assert dump_mode(ld, []) == 'ld'
     assert dump_mode(ld, [
         'foo.o', 'bar.o', 'baz.o', '-o', 'foo', '-Wl,-rpath,foo']) == 'ld'
 
 
-def test_ld_flags(wrapper_flags):
+def test_ld_flags(wrapper_environment, wrapper_flags):
     check_args(
         ld, test_args,
         ['ld'] +
@@ -204,7 +222,7 @@ def test_ld_flags(wrapper_flags):
         spack_ldlibs)
 
 
-def test_cpp_flags(wrapper_flags):
+def test_cpp_flags(wrapper_environment, wrapper_flags):
     check_args(
         cpp, test_args,
         ['cpp'] +
@@ -214,69 +232,58 @@ def test_cpp_flags(wrapper_flags):
         test_args_without_paths)
 
 
-def test_cc_flags(wrapper_flags):
+def test_cc_flags(wrapper_environment, wrapper_flags):
     check_args(
         cc, test_args,
         [real_cc] +
+        target_args +
         spack_cppflags +
         spack_cflags +
         spack_ldflags +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths +
+        common_compile_args +
         spack_ldlibs)
 
 
-def test_cxx_flags(wrapper_flags):
+def test_cxx_flags(wrapper_environment, wrapper_flags):
     check_args(
         cxx, test_args,
         [real_cc] +
+        target_args +
         spack_cppflags +
         spack_cxxflags +
         spack_ldflags +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths +
+        common_compile_args +
         spack_ldlibs)
 
 
-def test_fc_flags(wrapper_flags):
+def test_fc_flags(wrapper_environment, wrapper_flags):
     check_args(
         fc, test_args,
         [real_cc] +
+        target_args +
         spack_fflags +
         spack_cppflags +
         spack_ldflags +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths +
+        common_compile_args +
         spack_ldlibs)
 
 
-def test_dep_rpath():
+def test_dep_rpath(wrapper_environment):
     """Ensure RPATHs for root package are added."""
     check_args(
         cc, test_args,
         [real_cc] +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths)
+        target_args +
+        common_compile_args)
 
 
-def test_dep_include():
+def test_dep_include(wrapper_environment):
     """Ensure a single dependency include directory is added."""
     with set_env(SPACK_INCLUDE_DIRS='x'):
         check_args(
             cc, test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-Ix'] +
             test_library_paths +
@@ -285,7 +292,7 @@ def test_dep_include():
             test_args_without_paths)
 
 
-def test_system_path_cleanup():
+def test_system_path_cleanup(wrapper_environment):
     """Ensure SPACK_ENV_PATH is removed from PATH, even with trailing /
 
     The compiler wrapper has to ensure that it is not called nested
@@ -305,13 +312,14 @@ def test_system_path_cleanup():
             check_env_var(cc, 'PATH', system_path)
 
 
-def test_dep_lib():
+def test_dep_lib(wrapper_environment):
     """Ensure a single dependency RPATH is added."""
     with set_env(SPACK_LINK_DIRS='x',
                  SPACK_RPATH_DIRS='x'):
         check_args(
             cc, test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             test_library_paths +
             ['-Lx'] +
@@ -321,12 +329,13 @@ def test_dep_lib():
             test_args_without_paths)
 
 
-def test_dep_lib_no_rpath():
+def test_dep_lib_no_rpath(wrapper_environment):
     """Ensure a single dependency link flag is added with no dep RPATH."""
     with set_env(SPACK_LINK_DIRS='x'):
         check_args(
             cc, test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             test_library_paths +
             ['-Lx'] +
@@ -335,12 +344,13 @@ def test_dep_lib_no_rpath():
             test_args_without_paths)
 
 
-def test_dep_lib_no_lib():
+def test_dep_lib_no_lib(wrapper_environment):
     """Ensure a single dependency RPATH is added with no -L."""
     with set_env(SPACK_RPATH_DIRS='x'):
         check_args(
             cc, test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             test_library_paths +
             ['-Wl,--disable-new-dtags'] +
@@ -349,7 +359,7 @@ def test_dep_lib_no_lib():
             test_args_without_paths)
 
 
-def test_ccld_deps():
+def test_ccld_deps(wrapper_environment):
     """Ensure all flags are added in ccld mode."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_RPATH_DIRS='xlib:ylib:zlib',
@@ -357,6 +367,7 @@ def test_ccld_deps():
         check_args(
             cc, test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-Ixinc',
              '-Iyinc',
@@ -373,7 +384,7 @@ def test_ccld_deps():
             test_args_without_paths)
 
 
-def test_ccld_deps_isystem():
+def test_ccld_deps_isystem(wrapper_environment):
     """Ensure all flags are added in ccld mode.
        When a build uses -isystem, Spack should inject it's
        include paths using -isystem. Spack will insert these
@@ -386,6 +397,7 @@ def test_ccld_deps_isystem():
         check_args(
             cc, mytest_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-isystem', 'fooinc',
              '-isystem', 'xinc',
@@ -403,7 +415,7 @@ def test_ccld_deps_isystem():
             test_args_without_paths)
 
 
-def test_cc_deps():
+def test_cc_deps(wrapper_environment):
     """Ensure -L and RPATHs are not added in cc mode."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_RPATH_DIRS='xlib:ylib:zlib',
@@ -411,6 +423,7 @@ def test_cc_deps():
         check_args(
             cc, ['-c'] + test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-Ixinc',
              '-Iyinc',
@@ -420,7 +433,7 @@ def test_cc_deps():
             test_args_without_paths)
 
 
-def test_ccld_with_system_dirs():
+def test_ccld_with_system_dirs(wrapper_environment):
     """Ensure all flags are added in ccld mode."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_RPATH_DIRS='xlib:ylib:zlib',
@@ -434,6 +447,7 @@ def test_ccld_with_system_dirs():
         check_args(
             cc, sys_path_args + test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-Ixinc',
              '-Iyinc',
@@ -455,7 +469,7 @@ def test_ccld_with_system_dirs():
             test_args_without_paths)
 
 
-def test_ccld_with_system_dirs_isystem():
+def test_ccld_with_system_dirs_isystem(wrapper_environment):
     """Ensure all flags are added in ccld mode.
        Ensure that includes are in the proper
        place when a build uses -isystem, and uses
@@ -472,6 +486,7 @@ def test_ccld_with_system_dirs_isystem():
         check_args(
             cc, sys_path_args + test_args,
             [real_cc] +
+            target_args +
             test_include_paths +
             ['-isystem', 'xinc',
              '-isystem', 'yinc',
@@ -493,7 +508,7 @@ def test_ccld_with_system_dirs_isystem():
             test_args_without_paths)
 
 
-def test_ld_deps():
+def test_ld_deps(wrapper_environment):
     """Ensure no (extra) -I args or -Wl, are passed in ld mode."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_RPATH_DIRS='xlib:ylib:zlib',
@@ -514,7 +529,7 @@ def test_ld_deps():
             test_args_without_paths)
 
 
-def test_ld_deps_no_rpath():
+def test_ld_deps_no_rpath(wrapper_environment):
     """Ensure SPACK_LINK_DEPS controls -L for ld."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_LINK_DIRS='xlib:ylib:zlib'):
@@ -531,7 +546,7 @@ def test_ld_deps_no_rpath():
             test_args_without_paths)
 
 
-def test_ld_deps_no_link():
+def test_ld_deps_no_link(wrapper_environment):
     """Ensure SPACK_RPATH_DEPS controls -rpath for ld."""
     with set_env(SPACK_INCLUDE_DIRS='xinc:yinc:zinc',
                  SPACK_RPATH_DIRS='xlib:ylib:zlib'):
@@ -548,7 +563,7 @@ def test_ld_deps_no_link():
             test_args_without_paths)
 
 
-def test_ld_deps_partial():
+def test_ld_deps_partial(wrapper_environment):
     """Make sure ld -r (partial link) is handled correctly on OS's where it
        doesn't accept rpaths.
     """
@@ -586,57 +601,45 @@ def test_ld_deps_partial():
             test_args_without_paths)
 
 
-def test_ccache_prepend_for_cc():
+def test_ccache_prepend_for_cc(wrapper_environment):
     with set_env(SPACK_CCACHE_BINARY='ccache'):
         os.environ['SPACK_SHORT_SPEC'] = "foo@1.2=linux-x86_64"
         check_args(
             cc, test_args,
             ['ccache'] +  # ccache prepended in cc mode
             [real_cc] +
-            test_include_paths +
-            test_library_paths +
-            ['-Wl,--disable-new-dtags'] +
-            test_wl_rpaths +
-            test_args_without_paths)
+            target_args +
+            common_compile_args)
         os.environ['SPACK_SHORT_SPEC'] = "foo@1.2=darwin-x86_64"
         check_args(
             cc, test_args,
             ['ccache'] +  # ccache prepended in cc mode
             [real_cc] +
+            target_args +
             lheaderpad +
-            test_include_paths +
-            test_library_paths +
-            ['-Wl,--disable-new-dtags'] +
-            test_wl_rpaths +
-            test_args_without_paths)
+            common_compile_args)
 
 
-def test_no_ccache_prepend_for_fc():
+def test_no_ccache_prepend_for_fc(wrapper_environment):
     os.environ['SPACK_SHORT_SPEC'] = "foo@1.2=linux-x86_64"
     check_args(
         fc, test_args,
         # no ccache for Fortran
         [real_cc] +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths)
+        target_args +
+        common_compile_args)
     os.environ['SPACK_SHORT_SPEC'] = "foo@1.2=darwin-x86_64"
     check_args(
         fc, test_args,
         # no ccache for Fortran
         [real_cc] +
+        target_args +
         lheaderpad +
-        test_include_paths +
-        test_library_paths +
-        ['-Wl,--disable-new-dtags'] +
-        test_wl_rpaths +
-        test_args_without_paths)
+        common_compile_args)
 
 
 @pytest.mark.regression('9160')
-def test_disable_new_dtags(wrapper_flags):
+def test_disable_new_dtags(wrapper_environment, wrapper_flags):
     with set_env(SPACK_TEST_COMMAND='dump-args'):
         result = ld(*test_args, output=str).strip().split('\n')
         assert '--disable-new-dtags' in result
@@ -645,7 +648,7 @@ def test_disable_new_dtags(wrapper_flags):
 
 
 @pytest.mark.regression('9160')
-def test_filter_enable_new_dtags(wrapper_flags):
+def test_filter_enable_new_dtags(wrapper_environment, wrapper_flags):
     with set_env(SPACK_TEST_COMMAND='dump-args'):
         result = ld(*(test_args + ['--enable-new-dtags']), output=str)
         result = result.strip().split('\n')
@@ -657,7 +660,7 @@ def test_filter_enable_new_dtags(wrapper_flags):
 
 
 @pytest.mark.regression('22643')
-def test_linker_strips_loopopt(wrapper_flags):
+def test_linker_strips_loopopt(wrapper_environment, wrapper_flags):
     with set_env(SPACK_TEST_COMMAND='dump-args'):
         # ensure that -loopopt=0 is not present in ld mode
         result = ld(*(test_args + ["-loopopt=0"]), output=str)