Supersedes #26048, #25557.
Closes #25380.
Fixes #8703.

Third time's a charm?

This is a simpler take on converting `cc` to posix shell. The rationale for doing this can be found in https://github.com/spack/spack/pull/25380#issuecomment-903336139 and above.

#26048 worked, but the main issue there was that the code used a lot of command substitution with `$()`, which unfortunately runs in a subshell.  That requires a lot of forks, which are expensive, which is why we saw a lot of slowdowns with the POSIX wrappers in that PR.

## Design

This version tries to be as straightforward as possible.  Specifically, most conversions are kept simple -- convert ifs to ifs, handle indirect expansion the way we do in `setup-env.sh`, only mess with the logic in `cc`, and don't mess with the python code at all.

The big refactor is for arrays. We can't rely on bash's nice arrays and be ignorant of separators anymore. So:

1. To avoid complicated separator logic, there are three types of lists. They are:
    * `$lsep`-separated lists, which end with `_list`.  `lsep` is customizable, but we picked `^G` (alarm bell) for `$lsep` because it's ASCII and it's unlikely that it would actually appear in any arguments. If we need to get fancier (and I will lose faith in the world if we do) then we could consider XON or XOFF.
    * `:`-separated directory lists, which end with `_dirs`, `_DIRS`, `PATH`, or `PATHS`
    * whitespace-separated lists (like flags), which can have any other name
  
    Whitespace and colon-separated lists come with the territory with PATHs from env vars and lists of flags. `^G` separated lists are what we use for most internal variables, b/c it's more likely to work.
2. To avoid subshells, use a bunch of functions that do dirty `eval` stuff instead.  This adds 3 functions to deal with lists:
    * `append LISTNAME ELEMENT [SEP]`  will put `ELEMENT` at the end of the list called `LISTNAME`. You can optionally say what separator you expect to use. Note that we are taking advantage of everything being global and passing lists by name.
    * `prepend LISTNAME ELEMENT [SEP]`  like append, but puts `ELEMENT` at the start of `LISTNAME`
    *  `extend LISTNAME1 LISTNAME2 [PREFIX]` appends everything in LISTNAME2 to LISTNAME1, and optionally prepends `PREFIX` to every element (this is useful for things like `-I`, `-isystem `, etc.

The routines determine the separator for each argument by its name, so we don't have to pass around separators everywhere. Amazingly, as long as you do not expand variables' values within an `eval` environment, you can do all this and still preserve quoting. When iterating over lists, the user of this API still has to set and unset `IFS` properly.

We ended up having to ignore shellcheck SC2034 (unused variable), because using evals all over the place means that shellcheck doesn't notice that our list variables are actually used.

## Performance

So far this is looking pretty good.  I took the most complex unit test I could find to measure the performance of just the `cc` script.  It's here if you want to see it.

<details>
<summary>`test_cc.sh`: a moderately complex test script for `cc`</summary>
<p>

```shell
#!/bin/sh

export SPACK_TEST_COMMAND=dump-args

export SPACK_CC='/bin/mycc'
export SPACK_CXX='/bin/mycc'
export SPACK_FC='/bin/myfc'
export SPACK_PREFIX='/spack-test-prefix'
export SPACK_ENV_PATH='test'
export SPACK_DEBUG_LOG_DIR='.'
export SPACK_DEBUG_LOG_ID='foo-hashabc'
export SPACK_COMPILER_SPEC='gcc@4.4.7'
export SPACK_SHORT_SPEC='foo@1.2 arch=linux-rhel6-x86_64 /hashabc'
export SPACK_SYSTEM_DIRS='/bin:/usr/bin:/usr/local/bin:/bin64:/usr/bin64:/usr/local/bin64:/include:/usr/include:/usr/local/include:/lib:/usr/lib:/usr/local/lib:/lib64:/usr/lib64:/usr/local/lib64:/:/usr:/usr/local'
export SPACK_CC_RPATH_ARG='-Wl,-rpath,'
export SPACK_CXX_RPATH_ARG='-Wl,-rpath,'
export SPACK_F77_RPATH_ARG='-Wl,-rpath,'
export SPACK_FC_RPATH_ARG='-Wl,-rpath,'
export SPACK_LINK_DIRS=''
export SPACK_INCLUDE_DIRS=''
export SPACK_RPATH_DIRS=''
export SPACK_TARGET_ARGS=''
export SPACK_LINKER_ARG='-Wl,'
export SPACK_DTAGS_TO_ADD='--disable-new-dtags'
export SPACK_DTAGS_TO_STRIP='--enable-new-dtags'


export SPACK_CPPFLAGS='-g -O1 -DVAR=VALUE'
export SPACK_CFLAGS='-Wall'
export SPACK_CXXFLAGS='-Werror'
export SPACK_FFLAGS='-w'
export SPACK_LDFLAGS='-L foo'
export SPACK_LDLIBS='-lfoo'

env="/Users/gamblin2/src/spack/lib/spack/env"
cc="$env/cc"
ld="$env/ld"

$ld '-I/test/include' '-L/test/lib' '-L/other/lib' '-I/other/include' 'arg1' '-Wl,--start-group' 'arg2' '-Wl,-rpath,/first/rpath' 'arg3' '-Wl,-rpath' '-Wl,/second/rpath' '-llib1' '-llib2' 'arg4' '-Wl,--end-group' '-Xlinker' '-rpath' '-Xlinker' '/third/rpath' '-Xlinker' '-rpath' '-Xlinker' '/fourth/rpath' '-Wl,--rpath,/fifth/rpath' '-Wl,--rpath' '-Wl,/sixth/rpath' '-llib3' '-llib4' 'arg5' 'arg6' '-loopopt=0'
```

</p>
</details>

Using that, I can time ~200 invocations of cc like this:

```console
$ bash -c 'time (for i in `seq 1 200`; do ~/test_cc.sh > /dev/null; done)'
```

Using that, I get the following performance numbers (the listed shells are what I put in `cc`'s shebang):

**Original**
* Old version of `cc` with arrays and `bash v3.2.57` (macOS builtin): `4.462s` (`.022s` / call)
* Old version of `cc` with arrays and `bash v5.1.8` (Homebrew): `3.267s` (`.016s` / call)

**Using many subshells (#26408)**
*  with `bash v3.2.57`: `25.302s` (`.127s` / call)
*  with `bash v5.1.8`: `27.801s` (`.139s` / call)
*  with `dash`: `15.302s` (`.077s` / call)

This version didn't seem to work with zsh.

**This PR (no subshells)**
*  with `bash v3.2.57`: `4.973s` (`.025s` / call)
*  with `bash v5.1.8`: `4.984s` (`.025s` / call)
*  with `zsh`: `2.995s` (`.015s` / call)
*  with `dash`: `1.890s` (`.0095s` / call)        ⬅️ **Easily the winner** 🥇

So there are several interesting things to note here:

1. Running the posix version in `bash` is slower than using `bash` arrays.  That is to be expected because it's doing a bunch of string processing where it likely did not have to before, at least in `bash`.
2. `zsh`, at least on macOS, is significantly faster than the ancient `bash` they ship with the system. Using `zsh` with the new version also makes the posix wrappers faster than `develop`.  So it's worth preferring `zsh` if we have it.  I suppose we should also try this with newer `bash` on Linux.
3. `bash v5.1.8` seems to be significantly faster than the old system `bash v3.2.57` for arrays. For straight POSIX stuff, it's a little slower.  It did not seem to matter whether `--posix` was used.
4. `dash` is way faster than `bash` or `zsh`, so the real payoff just comes from being able to use it. I am not sure if that is mostly startup time, but it's significant. `dash` is ~2.4x faster than the original `bash` with arrays.

So I guess I may be proven wrong by @adamjstewart - I was right that doing a lot of string stuff is slower than arrays, but converting to posix seems worth it to be able to exploit `dash`.

- [x] Fix basic shellcheck issues
- [x] Convert arrays to use a few convenience functions: `append` and `extend`
- [x] Get `cc` tests passing.
- [x] More tests
- [x] Get this working with `gettext`'s `configure` and test on more packages.