@haampie 2021-09-27
> This should be fine, as we already pass paths to the compiler wrapper in colon-separated lists, and flags ultimately have to go to the compiler as space-separated argument lists. We may have to do something more to deal with quoting.

What do we have to escape? Can't we just follow how `PATH` works on unix? For instance on unix you can create funky dir names, but PATH does not support them:

```
$ mkdir \:
$ echo '#!/bin/sh'$'\n''echo hi' > \:/hello
$ chmod +x  \:/hello
$ \:/hello
hi

# literal?
$ export PATH="$(realpath \:):$PATH"
$ echo $PATH
/tmp/tmp.YWhxjbK6PX/::/usr/local/sbin: ...
$ hello # not found in path

# quoting?
$ export PATH='"/tmp/tmp.YWhxjbK6PX/:":'"$PATH"
$ echo $PATH
"/tmp/tmp.YWhxjbK6PX/:":/usr/local/sbin: ...
$ hello # not found in path

# escaping?
$ export PATH='/tmp/tmp.YWhxjbK6PX/\:'":$PATH"
$ /tmp/tmp.YWhxjbK6PX/\::/usr/local/sbin
$ hello # not found in path
```

There are no characters to quote/escape safely with, since `mkdir '\: ;"'` is a valid dir name too