@tgamblin 2021-09-28
Update: this takes about 20 seconds off the configure time for `gettext` on my laptop (if you change the shebang to point at dash).

228s vs 248s, or ~8% improvement for a real build, which is pretty cool in addition to the portability.