@tgamblin 2021-09-28
Ok I have a fix for the issue @haampie pointed out but it requires generating `cc` to remove the eval's (basically inlining them to avoid the quoting problems).  It works, and it preserves all quoting.  Worst case we can use that.  As we want to generate `cc` eventually anyway, this could work fine -- it will just make the actual generated `cc` less legible. 

Before I push that, though, @trws is working on a pure-sh escaping version based on the absolute insanity described here: https://stackoverflow.com/a/29824428/9122.  If that works, then we won't need to generate.