@tgamblin 2021-09-28
Ok turns out we don't actually need any of that fanciness to get this working.  @trws figured out that this was all we needed:

```diff
@@ -104,11 +104,11 @@ append() {
     elt="$2"

     if empty "$varname"; then
-        eval "$varname=\"${elt}\""
+        eval "$varname=\"\${elt}\""
     else
         # get appropriate separator for the list we're appending to
         setsep "$varname"
-        eval "$varname=\"\${$varname}${sep}${elt}\""
+        eval "$varname=\"\${$varname}${sep}\${elt}\""
     fi
 }
```

Two `\`'s is all it took.  The reason is that previously , `${elt}` was being expanded in the quoted string, so you would lose a quote level.  By deferring the evaluation to the `eval` environment, you don't ever lose the quote level because `${elt}` isn't ever expanded and passed to `eval`.

This is possible b/c `eval` is really only necessary to expand `$varname`, as posix sh doesn't so indirect expansion.  We don't *need* to pre-expand `${elt}` at all.

So the list interface here, with a little thought about `IFS`, seems to be pretty usable even in posix `sh`. Since we don't need to expand args -- just move them around -- we can preserve them, without messing with their quoting, through the whole script.

I also borrowed @cosmicexplorer's use of `^G` for the separator, since nobody's going to type a bell character. I just put a `^G` literal in the file to avoid a subshell.