@haampie 2021-09-28
The openblas / fftw bench on this PR:
|                                     | no compiler wrapper | develop        | #26048         | #26259         |
|-------------------------------------|---------------------|----------------|----------------|----------------|
| `openblas`                          | 53.94s, 54.10s      | 57.45s, 57.65s | 75.60s, 75.59s | 52.77s, 51.70s |
| `fftw`                              | 61.61s, 61.16s      | 63.80s, 64.21s | 70.95s, 71.98s | 60.43s, 59.65s |

It's so fast it's faster than no compiler wrapper :laughing: negative overhead, I like it!


```yaml
spack:
  config:
    ccache: false
    install_tree:
      root: ./here
  specs:
  - openblas
  - fftw ~mpi ~openmp precision=float
  view: false
  packages:
    perl:
      externals:
      - spec: perl@5.30.0~cpanm+shared+threads
        prefix: /usr
```