@tgamblin 2021-10-01
See above -- it looks like either @cosmicexplorer's editor or the GitHub suggestion API or something actually corrupts `^G`, which led to @haampie's error.  I noticed this only because I had seen before that multi-character `IFS` values caused all kinds of problems, which was how I discovered that `sh` treated `\a` as two characters.  Sigh.