@haampie 2021-10-02
On Alpine, tried this (had to pull in `python3-dev` & `git` for clingo, `linux-headers` for openssl)

```
apk add gcc g++ gfortran python3 python3-dev make patch bzip2 gzip linux-headers git
```

|            | no compiler wrapper | develop        | #26259         |
|------------|---------------------|----------------|----------------|
| `openblas` | 1m 19.97s      | 1m 22.35s | 1m 31.88s  |
| `fftw`     | 1m 24.29s      | 1m 26.30s | 1m 30.02s |