@tgamblin 2021-10-04
> Not sure what the story is here

Maybe that busy box is slower than bash?  At least it works though.  I'm assuming on `develop` you had to pull in bash, right?