@haampie 2021-10-04
> I'm assuming on develop you had to pull in bash, right?

yes!